LPD-72883 Create the CommentResourceTest class and migrate the comment tests.
